requirements.txt
boto3==1.35.19
botocore==1.35.19
html2text==2024.2.26
openai==1.45.1
pydantic==2.9.1
python_dateutil==2.9.0.post0
Requests==2.32.3
traitlets==5.14.3
